Change Servo process

+++ Code/Core/Inc/main.h
@@ -65,7 +65,7 @@ extern uint16_t Sensor_Threshold[8];
 extern uint8_t LineDectect;
 extern uint8_t GetThreshold_Flag;
 
-extern uint8_t ServoCountValue;
+extern uint16_t ServoCountValue;
 /* USER CODE END ET */
 
 /* Exported constants --------------------------------------------------------*/
